change module download->downloaders. Move github tests to their own test module
