+++ belay/packagemanager/__init__.py
@@ -1,1 +1,2 @@
+from . import downloaders
 from .group import Group, GroupConfig
