+++ belay/packagemanager/downloaders/common.py
