+++ tests/test_packagemanager.py
@@ -38,34 +38,6 @@ def main_group(tmp_path):
     )
 
 
-@pytest.mark.network
-def test_download_github_folder(tmp_path):
-    uri = "https://github.com/BrianPugh/belay/tree/main/tests/github_download_folder"
-    belay.packagemanager._download_github(tmp_path, uri)
-
-    assert (tmp_path / "__init__.py").exists()
-    assert (
-        tmp_path / "file1.py"
-    ).read_text() == 'print("belay test file for downloading.")\n'
-    assert (
-        tmp_path / "file2.txt"
-    ).read_text() == "File for testing non-python downloads.\n"
-    assert (tmp_path / "submodule" / "__init__.py").exists()
-    assert (
-        tmp_path / "submodule" / "sub1.py"
-    ).read_text() == 'foo = "testing recursive download abilities."\n'
-
-
-@pytest.mark.network
-def test_download_github_single(tmp_path):
-    uri = "https://github.com/BrianPugh/belay/blob/main/tests/github_download_folder/file1.py"
-    belay.packagemanager._download_github(tmp_path, uri)
-
-    assert (
-        tmp_path / "__init__.py"
-    ).read_text() == 'print("belay test file for downloading.")\n'
-
-
 def test_download_all(main_group, mocker, spy_ast):
     main_group.download()
